+++ requirements.txt
@@ -18,3 +18,4 @@ pytest
 torch
 torch-tb-profiler
 tensorboard
+aiosqlite
